# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Architecture Knowledge Platform API
  version: 0.2.0
  x-akp-write-idempotency: >
    Mutating POST operations accept Idempotency-Key and MCP write tools require
    it. POST /v1/auth/session is intentionally exempt because it creates a
    Set-Cookie response that cannot be safely replayed.
  description: >
    HTTP contract for the routes registered by the current Fastify runtime.
    All /v1 operations require a scoped bearer API token or an unexpired
    HttpOnly web session and enforce the permission recorded in
    x-akp-permission. Session-authenticated writes also require X-CSRF-Token.
servers:
  - url: http://127.0.0.1:8080
    description: Local development default
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /v1/status:
    get:
      tags: [Knowledge]
      operationId: getPlatformStatus
      x-akp-permission: knowledge:read
      responses:
        "200":
          description: Authorized corpus and capability status
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/context-packs/{id}:
    get:
      tags: [Knowledge]
      operationId: getContextPack
      x-akp-permission: knowledge:read
      parameters:
        - { $ref: "#/components/parameters/StableId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/documents/{id}:
    get:
      tags: [Knowledge]
      operationId: getDocument
      x-akp-permission: knowledge:read
      parameters:
        - { $ref: "#/components/parameters/StableId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/documents/{id}/evidence:
    get:
      tags: [Knowledge]
      operationId: getDocumentEvidence
      x-akp-permission: source:read
      parameters:
        - { $ref: "#/components/parameters/StableId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/impact/{id}:
    get:
      tags: [Knowledge]
      operationId: analyzeKnowledgeImpact
      x-akp-permission: knowledge:read
      parameters:
        - { $ref: "#/components/parameters/StableId" }
        - in: query
          name: depth
          schema: { type: integer, minimum: 1, maximum: 5, default: 2 }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vaults:
    get:
      tags: [Knowledge]
      operationId: listVaults
      x-akp-permission: knowledge:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Knowledge, Governance]
      operationId: registerVault
      x-akp-permission: admin
      x-akp-requires-unrestricted-space: true
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VaultRegistration" }
      responses:
        "201": { $ref: "#/components/responses/ObjectResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/graph/summary:
    get:
      tags: [Knowledge]
      operationId: getGraphSummary
      x-akp-permission: knowledge:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/sources:
    get:
      tags: [Sources]
      operationId: listSources
      x-akp-permission: source:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/sources/{id}:
    get:
      tags: [Sources]
      operationId: getSource
      x-akp-permission: source:read
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: >
        Provisioned out of band with an explicit TokenScopeSet. Token scopes
        are intersected with the user's current memberships on every request.
    cookieAuth:
      type: apiKey
      in: cookie
      name: akp_session
      description: HttpOnly session carrying the effective TokenScopeSet snapshot.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Actor- and operation-scoped key. Reuse with a different request returns 409.
      schema: { type: string, minLength: 8, maxLength: 200 }
...
    UuidId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    StableId:
      in: path
      name: id
      required: true
      description: Database UUID, stable external ID, alias or title where supported.
      schema: { type: string, minLength: 1 }
  responses:
    ObjectResponse:
      description: Runtime response object
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, expired, revoked or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Permission, space or path scope denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found in the authorized scope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Resource state, publication lock or optimistic revision conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
      additionalProperties: true
...
    VaultEvalPack:
      type: object
      additionalProperties: false
      required: [name, version]
      properties:
        name:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
        version: { type: string, minLength: 1 }
        enabled: { type: boolean, default: true }
        criticalCases:
          type: array
          default: []
          items: { type: string }
    VaultRegistration:
      type: object
      additionalProperties: false
      required: [vaultKey, name, spaceId, localPath]
      properties:
        vaultKey:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
        name: { type: string, minLength: 1, maxLength: 200 }
        spaceId: { type: string, format: uuid }
        visibility:
          type: string
          enum: [PRIVATE, TEAM, CENTRAL]
          default: PRIVATE
        gitRepository: { type: [string, "null"] }
        defaultBranch: { type: string, minLength: 1, default: main }
        localPath: { type: string, minLength: 1 }
        contentRoots:
          type: array
          minItems: 1
          default: ["."]
          items: { type: string, minLength: 1 }
        sourceRoots:
          type: array
          default: []
          items: { type: string, minLength: 1 }
        schemaProfile: { type: object, additionalProperties: true, default: {} }
        evalPack: { $ref: "#/components/schemas/VaultEvalPack" }
        retrievalConfig:
          { type: object, additionalProperties: true, default: {} }
        permissions: { type: object, additionalProperties: true, default: {} }
        enabled: { type: boolean, default: true }
...